+++ packages/kit/src/core/config/index.spec.js
@@ -98,10 +98,7 @@ const get_defaults = (prefix = '') => ({
 		host: undefined,
 		hydrate: undefined,
 		inlineStyleThreshold: 0,
-		methodOverride: {
-			parameter: '_method',
-			allowed: []
-		},
+		methodOverride: undefined,
 		moduleExtensions: ['.js', '.ts'],
 		outDir: join(prefix, '.svelte-kit'),
 		serviceWorker: {