+++ packages/kit/src/core/sync/write_types/test/simple-page-server-only/+page.server.js
@@ -3,3 +3,7 @@ export function load() {
 		foo: 'bar'
 	};
 }
+
+export const actions = {
+	default: () => ({ foo: 'bar' })
+};