+++ packages/kit/test/apps/basics/src/routes/actions/success-data/+page.server.js
@@ -1,0 +1,16 @@
+/** @type {import('./$types').PageServerLoad} */
+export function load() {
+	return {
+		initial: 'initial'
+	};
+}
+
+/** @type {import('./$types').Actions} */
+export const actions = {
+	default: async ({ request }) => {
+		const fields = await request.formData();
+		return {
+			result: fields.get('username')
+		};
+	}
+};