+++ packages/kit/test/apps/basics/src/routes/prerendering/mutative-endpoint/+page.server.ts
@@ -1,3 +1,3 @@
-export const POST = () => {
+export const actions = () => {
 	return {};
 };