+++ packages/kit/test/apps/basics/src/routes/shadowed/no-get/+page.server.js
@@ -1,3 +1,6 @@
-export function POST() {
-	return {};
-}
+/** @type {import('./$types').Actions} */
+export const actions = {
+	default: () => {
+		return {};
+	}
+};