+++ packages/kit/types/internal.d.ts
@@ -1,9 +1,7 @@
 import { OutputAsset, OutputChunk } from 'rollup';
 import { SvelteComponent } from 'svelte/internal';
 import {
-	Action,
 	Config,
-	ExternalFetch,
 	ServerLoad,
 	Handle,
 	HandleServerError,
@@ -15,6 +13,8 @@ import {
 	Server,
 	ServerInitOptions,
 	SSRManifest,
+	HandleFetch,
+	Actions,
 	HandleClientError
 } from './index.js';
 import {
@@ -91,7 +91,7 @@ export type CSRRoute = {
 export type GetParams = (match: RegExpExecArray) => Record<string, string>;
 
 export interface ServerHooks {
-	externalFetch: ExternalFetch;
+	handleFetch: HandleFetch;
 	handle: Handle;
 	handleError: HandleServerError;
 }
@@ -120,11 +120,6 @@ export interface ManifestData {
 	nodes: PageNode[];
 	routes: RouteData[];
 	matchers: Record<string, string>;
-}
-
-export interface MethodOverride {
-	parameter: string;
-	allowed: string[];
 }
 
 export interface PageNode {
@@ -283,10 +278,7 @@ export interface SSRNode {
 		prerender?: PrerenderOption;
 		ssr?: boolean;
 		csr?: boolean;
-		POST?: Action;
-		PATCH?: Action;
-		PUT?: Action;
-		DELETE?: Action;
+		actions?: Actions;
 	};
 
 	// store this in dev so we can print serialization errors
@@ -304,7 +296,6 @@ export interface SSROptions {
 	handle_error(error: Error & { frame?: string }, event: RequestEvent): App.PageError;
 	hooks: ServerHooks;
 	manifest: SSRManifest;
-	method_override: MethodOverride;
 	paths: {
 		base: string;
 		assets: string;